add token in server

+++ BankFinalProjectServer/Account.Data/Entities/Account.cs
@@ -10,5 +10,5 @@ public class Account
     [Required]
     public DateTime OpenDate { get; set; }
     [Required]
-    public decimal Balance { get; set; } = 1000;
+    public int Balance { get; set; } = 1000;
 }
